PATH bug fixed
