+++ .vscode/configurationCache.log
@@ -1,0 +1,1 @@
+{"buildTargets":["all","clean","fclean","libft/libft.a","re"],"launchTargets":["/Users/ilahyani/Minishell>minishell()"],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":["/Users/ilahyani/Minishell","/opt/homebrew/opt/readline/include"],"compilerArgs":["-Wall","-Wextra","-Werror","-lreadline","-L/opt/homebrew/opt/readline/lib","./libft/libft.a","env_init.c","execve.c","expantion.c","main.c","my_cd.c","my_echo.c","my_env.c","my_exit.c","my_export.c","my_pwd.c","my_unset.c","sig_handle.c","utils.c","-o","minishell"],"compilerPath":"/usr/bin/cc","standard":"c11","windowsSdkVersion":""},"fileIndex":[["/Users/ilahyani/Minishell/*.c",{"uri":{"$mid":1,"fsPath":"/Users/ilahyani/Minishell/*.c","path":"/Users/ilahyani/Minishell/*.c","scheme":"file"},"configuration":{"defines":[],"standard":"c11","includePath":["/opt/homebrew/opt/readline/include"],"forcedInclude":[],"intelliSenseMode":"clang-x64","compilerPath":"/usr/bin/cc","compilerArgs":["-Wall","-Wextra","-Werror","-lreadline","-L/opt/homebrew/opt/readline/lib","./libft/libft.a","env_init.c","execve.c","expantion.c","main.c","my_cd.c","my_echo.c","my_env.c","my_exit.c","my_export.c","my_pwd.c","my_unset.c","sig_handle.c","utils.c","-o","minishell"],"windowsSdkVersion":""},"compileCommand":{"command":"cc -Wall -Wextra -Werror -lreadline -I/opt/homebrew/opt/readline/include -L/opt/homebrew/opt/readline/lib ./libft/libft.a *.c -o minishell","directory":"/Users/ilahyani/Minishell","file":"/Users/ilahyani/Minishell/*.c"}}]]}}
