+++ main.c
@@ -6,7 +6,7 @@
 /*   By: ilahyani <ilahyani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 10:28:26 by ilahyani          #+#    #+#             */
-/*   Updated: 2022/07/09 17:51:26 by ilahyani         ###   ########.fr       */
+/*   Updated: 2022/07/12 23:17:45 by ilahyani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,88 +20,16 @@ void    check_cmd(char **cmd, t_env *lst_env, t_env *expand)
         g_exit = my_cd(lst_env, cmd);
     else if (!ft_strcmp(cmd[0], "echo"))
         g_exit = my_echo(cmd, lst_env, expand);
-    else if (!ft_strcmp(cmd[0], "exit"))
-        my_exit(cmd);
     else if (!ft_strcmp(cmd[0], "export"))
         g_exit = my_export(cmd, lst_env);
     else if (!ft_strcmp(cmd[0], "env"))
         g_exit = my_env(lst_env, cmd);
     else if (!ft_strcmp(cmd[0], "unset"))
         g_exit = my_unset(lst_env, cmd);
+    else if (!ft_strcmp(cmd[0], "exit"))
+        my_exit(cmd);
     else
         g_exit = ft_exec(cmd, lst_env);
-}
-
-int check_arg(char **args)
-{
-    int i;
-
-    i = -1;
-    while (find_char(args[++i], '='));
-    return (i);
-}
-
-t_env   *add_node(char *buff, t_env *expantion)
-{
-    char    *key;
-    char    *value;
-    key = ft_strldup(buff, find_char(buff, '=')); //free
-    if (ft_strlen(strchr_plus(buff, '=')) > 0)
-    {
-        value = ft_strdup(strchr_plus(buff, '=')); //free
-        env_lstadd_back(&expantion, env_lstnew(key, value));
-    }
-    else
-        env_lstadd_back(&expantion, env_lstnew(key, ""));
-    printf("var1 = %s\n",expantion->var);
-    printf("value1 = %s\n",expantion->value);
-    return expantion;
-}
-
-t_env   *update_node(char *buff, t_env *expantion)
-{
-    if (ft_strlen(strchr_plus(buff, '=')) > 0)
-        expantion->value = ft_strdup(strchr_plus(buff, '='));
-    else
-        expantion->value = ft_strdup("");
-    return (expantion);
-}
-
-t_env   *check_expantion(char **args, t_env *lst_env, t_env *expantion)
-{
-    t_env   *tmp;
-    int     i;
-
-    if (check_arg(args) == sizeof_array(args))
-    {
-        i = -1;
-        while (args[++i])
-        {
-            tmp = expantion;
-            while (tmp && ft_strcmp(tmp->var, ft_strldup(args[i], find_char(args[i], '='))))
-                tmp = tmp->next;
-            if (!tmp)
-                expantion = add_node(args[i], expantion);
-            else
-                expantion = update_node(args[i], expantion);
-        }
-    }
-    else
-        check_cmd(args + check_arg(args), lst_env, expantion);
-    return (expantion);
-}
-
-void    handler(int signum)
-{
-    if (signum == SIGINT)
-    {
-        write(1, "\n", 1);
-		rl_replace_line("", 0);
-		rl_on_new_line();
-		rl_redisplay();
-    }
-    else if (signum == SIGQUIT)
-        return ;
 }
 
 void    ft_readline(t_env *lst_env, t_env *expand)
@@ -111,7 +39,7 @@ void    ft_readline(t_env *lst_env, t_env *expand)
 
     while (1)
     {
-        line = readline("🐌🐌🐌 ~ ");
+        line = readline("🐌🐌🐌 ~> ");
         if (!line)
         {
             g_exit = 0;
@@ -127,7 +55,8 @@ void    ft_readline(t_env *lst_env, t_env *expand)
             expand = check_expantion(data, lst_env, expand);
         else
             check_cmd(data, lst_env, expand);
-        add_history(line);
+        if (ft_strlen(line) > 0)
+				add_history(line);
         //free data
     }
 }
@@ -143,10 +72,7 @@ int main(int ac, char **av, char **env)
     expand = NULL;
     signal(SIGINT, handler);
     signal(SIGQUIT, handler);
-    // if (!env) manually set shlvl oldpwd(empty) pwd + _ (doesn't show in export)
-    lst_env = env_init(env, lst_env);
-    //send address instead of value here and above + manage lists in a seperate file
-    //unset oldpwd & export oldpwd(empty) & export pwd=getcwd
+    env_init(env, &lst_env);
     ft_readline(lst_env, expand);
     return (0);
 }
