+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: ilahyani <ilahyani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 10:28:22 by ilahyani          #+#    #+#             */
-/*   Updated: 2022/07/01 16:45:16 by ilahyani         ###   ########.fr       */
+/*   Updated: 2022/07/11 16:47:25 by ilahyani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,37 +54,6 @@ char	**parser(char *line)
 	return (data);
 }
 
-char	**sort_tab(char **tab)
-{
-	char	**sorted;
-	char	*tmp;
-	int		i;
-	int		j;
-
-	sorted = (char **) malloc(sizeof(char *) * (sizeof_array(tab) + 1));
-	i = -1;
-	while (tab[++i])
-		sorted[i] = ft_strdup(tab[i]);
-	sorted[i] = NULL;
-	i = 0;
-    while (sorted[i])
-    {
-		j = i + 1;
-        while (sorted[j])
-        {
-            if (ft_strcmp(sorted[i], sorted[j]) > 0)
-            {
-				tmp = sorted[i];
-                sorted[i] = sorted[j];
-                sorted[j] = tmp;
-            }
-            j++;
-        }
-        i++;
-    }
-	return (sorted);
-}
-
 char	*strchr_plus(const char *s, int c)
 {
 	size_t	i;
@@ -115,64 +84,10 @@ int	find_char(char *s, char c)
 	return (0);
 }
 
-t_env	*env_lstlast(t_env *lst)
+void	err_print(char *cmd, char *buff)
 {
-	t_env	*last;
-
-	if (!lst)
-		return (0);
-	last = lst;
-	while (last->next)
-		last = last->next;
-	return (last);
-}
-
-void	env_lstadd_back(t_env **lst, t_env *new)
-{
-	t_env	*last;
-
-	if (!new)
-		return ;
-	if (!*lst)
-	{
-		*lst = new;
-		return ;
-	}
-	last = env_lstlast(*lst);
-	last->next = new;
-}
-
-t_env	*env_lstnew(char *var, char *value)
-{
-	t_env	*res;
-
-	res = (t_env *) malloc (sizeof(t_env));
-	if (!res)
-		return (0);
-	res->var = ft_strdup(var);
-    if (value)
-        res->value = ft_strdup(value);
-    else
-        res->value = NULL;
-	res->next = NULL;
-	return (res);
-}
-
-t_env    *env_init(char **env, t_env *lst_env)
-{
-    char **buff;
-    int i;
-
-    i = -1;
-    while (env[++i])
-    {
-        buff = ft_split(env[i], '=');
-        if (!buff[1])
-            buff[1] = NULL;
-	    env_lstadd_back(&lst_env, env_lstnew(buff[0], buff[1]));
-        free(buff[0]);
-        free(buff[1]);
-        free(buff);
-    }
-    return (lst_env);
+	ft_putstr_fd("minishell: ", 2);
+	ft_putstr_fd(cmd, 2);
+    ft_putstr_fd(": ", 2);
+	ft_putendl_fd(buff, 2);
 }
